backend/requirements.txt
flask
pyjwt>=2.8.0
sendgrid
python-dotenv
requests
psutil
Flask-SocketIO
Flask-CORS
python-socketio
python-engineio
Werkzeug
